+++ server/src/main/java/com/yeohaeng_ttukttak/server/oauth2/service/provider/apple/AppleClientSecretProvider.java
@@ -1,6 +1,6 @@
 package com.yeohaeng_ttukttak.server.oauth2.service.provider.apple;
 
-import com.yeohaeng_ttukttak.server.token.TokenService;
+import com.yeohaeng_ttukttak.server.token.provider.JwtProvidable;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -16,7 +16,7 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class AppleClientSecretProvider {
 
-    private final TokenService tokenService;
+    private final JwtProvidable jwtProvidable;
     private final AppleOAuthProps props;
 
     private String clientSecret;
@@ -24,7 +24,7 @@ public class AppleClientSecretProvider {
     @Scheduled(fixedRate = 14, timeUnit = TimeUnit.DAYS)
     protected void renew() {
 
-        this.clientSecret = tokenService.issueByES256(
+        this.clientSecret = jwtProvidable.issueByES256(
                 props.privateKey(),
                 Duration.ofDays(21),
                 Map.of("kid", props.keyId()),
