--- server/src/main/java/com/yeohaeng_ttukttak/server/token/dto/IdTokenClaim.java
@@ -1,6 +1,0 @@
-package com.yeohaeng_ttukttak.server.token.dto;
-
-public record IdTokenClaim(
-        String openId,
-        String name
-) { }
